Started local version of adapthisteq that ignores masked pixels
Annoyingly, the MATLAB adapthisteq function makes use of several MEX files that
are stored in the image processing toolbox's "private" directory. The easiest
solution is to make a copy of these files in a local private directory. There
are sevearl reasons this is not the best solution (the mex files are platform
specific, copyright/distribution issues are unclear, ...) but the alternative
(to re-implement the algorithm in part or in whole) are not worth the time to me
right now. This decision will have to be revisited if I do decide to distribute
this program.

See here for more about private directories:
http://www.mathworks.com/help/matlab/matlab_prog/private-functions.html