Cleaned up treatment of private files needed by intensity
The mex files turn out to be version dependant. There is no "solution" for this
issue, but it is possible to manage it more cleanly. I do this by merging the
masked_adapthisteq function into yalebox_prep_intensity, and reserving the
private folder for the mex files only. Rather than copying these, I create
symbolic links to them in private. Lastly, MATLAB changed the name of one mex
dependency between 2014a and 2014b, to deal with this I alias the name of the
older version to the newer version if needed.