step: updating grid to use linspace to span the domain
The old approach was to select a fixed grid spacing for each pass, which had the
annoying consequence that the grid points typically did not span the domain. The
result was a complicated set of interpolation/extrapolation steps needed to
expand the convex hull of each pass to cover all the points in the new pass.

This new approach instead accepts a number of grid points in both dimensions as
an input, and uses the linspace() function to compute grid points that span the
domain.

The advantage is that all passes will have the same footprint, and so
the awkward extrapolation steps can be omitted.

One disadvantage is that the cell centers do not in general lie at integer
positions in pixel coordinates.  this means that they will have to be rounded,
and introduces an uncertainty of < 1 pixel into the x- and y- position of the
outputs. I consider this to entirely negligable.

Another disadvantage is that it may be quite tricky to define an equally-spaced
grid (or at least approximately equally spaced). My solution to this is to allow
the user to supply an empty value (0) for one of the two resolutionss, and to
compute a value that approximately matches the spacing in pixel coordinates.
This is done in the parse_inputs() function.