step: changed mask/roi test
Previously, a point had to have a sample window that contained a more than a
minimum fraction of data (i.e. not mask).

Now, a point must have a center that lies within the data.

This test is simpler, more efficient, and requires no parameters. A cursory
comparison indicates that it does not degrade the results either.