Testing different intepolation/extrapolation methods
Generating a deformation field requires interpolation and extrapolation. I
wondered if the interp2 'spline' method was the source of the poor performance
at the edge. It turns out that the griddedInterpolant class allows more
flexibility in extrapolation (and should be faster too), so I switched to this
method and tried two alternatives, 'linear' and 'spline'. The difference turns
out to be negligible, with a slight advantage for spline.

LINEAR:

displacement vector error quantiles
qnt | uu    | vv    | mag
0.00  | 1.43e-05  | 6.41e-05  | 8.81e-04
0.10  | 2.54e-03  | 1.38e-03  | 5.42e-03
0.20  | 4.34e-03  | 2.27e-03  | 9.02e-03
0.30  | 7.14e-03  | 3.28e-03  | 1.12e-02
0.40  | 1.10e-02  | 4.86e-03  | 1.41e-02
0.50  | 1.43e-02  | 6.21e-03  | 1.75e-02
0.60  | 1.95e-02  | 7.39e-03  | 2.26e-02
0.70  | 2.88e-02  | 9.28e-03  | 3.16e-02
0.80  | 4.92e-02  | 1.21e-02  | 4.97e-02
0.90  | 3.17e-01  | 1.99e-02  | 3.18e-01
1.00  | 4.63e-01  | 5.04e-02  | 4.64e-01

SPLINE:

displacement vector error quantiles
qnt | uu    | vv    | mag
0.00  | 1.08e-05  | 3.46e-06  | 7.09e-04
0.10  | 2.11e-03  | 1.27e-03  | 4.86e-03
0.20  | 4.14e-03  | 2.16e-03  | 8.21e-03
0.30  | 6.26e-03  | 3.27e-03  | 1.07e-02
0.40  | 9.55e-03  | 4.72e-03  | 1.27e-02
0.50  | 1.27e-02  | 6.11e-03  | 1.56e-02
0.60  | 1.79e-02  | 7.58e-03  | 2.03e-02
0.70  | 2.54e-02  | 9.45e-03  | 2.77e-02
0.80  | 3.77e-02  | 1.19e-02  | 4.16e-02
0.90  | 2.64e-01  | 1.93e-02  | 2.65e-01
1.00  | 4.68e-01  | 7.60e-02  | 4.68e-01