Working grid refinement implementation, some issues in testing
After grid refinement, there are some really nasty edge artefacts in the
constant_offset test case. Looking at the correlation planes shows that the zero
edges are dominating the signal. Perhaps CBC would help in this case, and simply
was not needed in the previous example which had larger sample windows?