Working lanzos optimization routine, not robust
The results look comparable to gauss2d for the constant offset test (see below).
The method is not yet robust to peaks that lie within nkern+1 points of the edge
of the correlation plane. I think the best way to deal with this is to reject
such peaks, although an alternative would be to add a symmetric pad.

OPTIM_LANCZOS:

qnt | uu    | vv    | mag
0.00  | 1.69e-05  | 1.35e-06  | 1.99e-05
0.10  | 1.21e-04  | 7.31e-05  | 1.85e-04
0.20  | 2.23e-04  | 1.45e-04  | 2.93e-04
0.30  | 3.84e-04  | 2.24e-04  | 5.26e-04
0.40  | 5.27e-04  | 3.15e-04  | 8.26e-04
0.50  | 7.22e-04  | 3.87e-04  | 1.19e-03
0.60  | 1.01e-03  | 5.32e-04  | 1.54e-03
0.70  | 1.41e-03  | 8.45e-04  | 1.94e-03
0.80  | 1.95e-03  | 1.23e-03  | 2.29e-03
0.90  | 2.53e-03  | 1.61e-03  | 2.75e-03
1.00  | 4.33e-03  | 2.95e-03  | 5.24e-03

GAUSS2D:

0.00  | 2.05e-04  | 4.18e-04  | 4.66e-04
0.10  | 2.05e-04  | 4.18e-04  | 4.66e-04
0.20  | 2.05e-04  | 4.18e-04  | 4.66e-04
0.30  | 2.05e-04  | 4.18e-04  | 4.66e-04
0.40  | 2.05e-04  | 4.18e-04  | 4.66e-04
0.50  | 2.05e-04  | 4.18e-04  | 4.66e-04
0.60  | 2.05e-04  | 4.18e-04  | 4.66e-04
0.70  | 2.05e-04  | 4.18e-04  | 4.66e-04
0.80  | 2.05e-04  | 4.18e-04  | 4.66e-04
0.90  | 2.05e-04  | 4.18e-04  | 4.66e-04
1.00  | 2.05e-04  | 4.18e-04  | 4.66e-04