Updated morphological filtering in automatic masking routine
The new approach uses an opening filter to remove noise outside of the wedge and
to simplify the wedge boundary, then an erosion to trim off the "halo" of
non-wedge pixels that were previously included in the mask. A larger strel for
the erosion could potentially also remove bright edges at the wedge surface.