Starting new experiment
This experiment explores the effect of accounting for the offset between the
sample window center and the data (a.k.a. sand) for sample windows that overlap
the wedge boundary. This offset means that the sample grid is not in fact
regular. To populate the regular grid, it will be necessary to identify the
centroid of boundary points, then interpolate the data from the scattered grid
to the desired regular grid.