ahe: complete global_he.m, a simple global histogram equalization routine
The equalized image has a linear CDF and a flat PDF. Note that a histogram with
arbitrary (including equally spaced) bins will NOT be flat since the input data
are in fact discrete.

This discrete data is a reason for concern --- there are only 155 unique values
in the masked input data. I wonder if this is inherent to the dataset, or if it
can be improved by using a different method to convert to grayscale.