ahe: Completed brute force local equalization routine
The implementation is very inefficient, and needs to be run in parallel to
complete in a reasonable time.

Results show a visually satisfactory equalized image with almost no long
wavelength features remaining. The equalized image CDF appears linear at
low-zoom, but in detail it is a stair-step function, which means the PDF (its
derivative) is a mess.

The steps are likely caused by the fact that the local neighborhood kernel is
unweighted - so pixels pop-in and pop-out for each discrete shift of the window.
It is not clear that this is a problem in terms of the results, and the effect
may be reduced in a more efficient implementation.