ahe: abandoning attempts at a more efficient implementation
Dealing with the boundary of the region of interest is a little tricky, and this
step is taking more time than it is worth. Brute-force it is for now.

A few notes before moving on:

One interesting approach to the boundary issue is to compute the transformations
on an irregular grid and use scattered interpolation to map each pixel. A
suitable mesh can be generated using the DistMesh tool by Per-Oolof Persson
(http://persson.berkeley.edu/distmesh/index.html) using a simplified version of
the ROI boundary for the "distance function". Of course, the cost of generating
a decent mesh makes this approach as slow as brute force!