fixed bug in sample grid generation
Bug was introduced the with 'experimental' grid expansion. Some sample windows
were completely empty, and the rounding/padding routines in window extraction
did not work properly in these cases. It makes no sense to generate a grid where
whole rows or columns will be always be skipped, so I rewrote the grid
generation routine. In the new grid, the footprint of the sample windows is
always larger than the image, and all sample windows on the boundary contain
some data (how much depends on the grid parameters and image size).