synth_bnd: generates ini image with gaussian particle images...
This works, but it is not clear that it is quite what I want. The particle
images are either entirely separate (for small sigma), as in a true PIV
experiment, or completely overlapping (for large sigma) with values >> 1.

The latter case is more like the sandbox, but perhaps more accurate would be to
assign some points negative amplitudes in order to replicate teh black-and-white
sand particles.