debug/centroid: centroid bug fixed, problem not solved.
Debugging plots indicate that the program correctly identifies the centroid of
the sample window. However, errors remain larger when the centroid grid is used
(errors are smaller if the sample grid is NOT adjusted to reflect the data
centroids).

I suspect that the cause is that interpolation from the centroid grid (with
missing values) to the full sample grid is computed on unsmoothed data, and
propagates local noise into the interpolated points.

Incorporating smoothing into the interpolation routine (as suggested by Wessel &
Bercovici) may avoid propagating errors during interpolation. In addition, the
current smoothing routine (PLS) is known to perform poorly at data boundaries,
and needs to be replaced.