edge_gradient: cleaned up version for merge...
Current state of the code is that to use the spline-in-tension interpolation
routine for both sample and full grid. This yeilds error maps with no noticable
edge gradients, but strain values still have a gradient.

Next steps are to look for bugs in the gradient/strain routine, and to figure
out how to add smoothing in the multipass routine.