First implementation using simple kernel smoother...
I use a convolution routine from the File Exchange that ignores NaNs (by
normalizing the ouput with the convolution of a mask matrix).  Errors decrease,
but edge gradients appear. On inspection, it is clear that this behavior is
correct --- in the presence of a gradient, there is a bias introduced by the
fact that the kernel is not symmetric about the center point at the boundaries.

One solution might be to smooth the corrector values during each pass, since
these should converge to a noisy zero after several passes.