Multipass multi-resolution working...
Results are somewhat disappointing, in that errors increase at higher
resolutions. I suspect that a smoothing step is necessary.

Also, edge gradients appear for small window sizes. These are eliminated if the
min_data_frac constant in piv_displacements is increased. Most likely, there is
not enough data in a half-empty small window.