piv_series: attempt to correct the dimensions of the output file...
These are NOT equal to the input file, but rather are derived from the image
dimensions and the sample grid spacing. This initial attempt stumbled on a
precision problem (the data are initially read in as single precision).

I plan to switch to a more elegant solution in which the piv_sample_grid
routine also returns the world coordinates.