movie_displacement: abandoned internal VideoWriter for image frames...
New appraoch is to generate image frames at the desired resolution (FHD), then
stich these together into a video using the full-featured avconv external
program. The motivation is the difficulty of specifying 1) video format, and 2)
video resolution. Specifically, MATLAB does not provide many options for video
encoding on Linux, and video playback does not perform well for arbitrary
resolutions. The solution is to use MATLAB to generate images (which is hard
enough) and then create a video externally.