piv_check_manual: added button to delete points
This is needed because some regions are difficult or impossible to find matches
in. The addition also required adding an argument to the next_pt function to
make recording the previous point options. Without this, the record_pt routine
gets the control point location from the impoint objects, effectively re-adding
the point that is to be deleted.