Exploring alternative approach to computing F
I computed the original and final coordinates and computed F directly. This
works, except that nearest-neighbor padding breaks (there *should* be a
gradient in coordinates as opposed to displacements).

The next step is to go back to the original approach of differentiating the
displacements, but adjust the coordinates to make the calculation correct.