Added explicit handling for common warning
It is common for the overlap between the ROIs of the interrogation window and
sample window to be everywhere less than the minimum. For example at the left
and right sides of the image, warping shifts the ROI mask and can reduce
overlap for later iterations. This is expected behavior, and should not throw a
warning.