Experimented with new approach to modeling errors in test script
When struggling with how to "window" the data when producing a "functional
box plot", I realized that it is much more straight forward to discuss the
conditional CDF P(error | distance to shear center). This can be estimated
with kernels, with the usual caveats about bandwidth. I found a reasonable
adaptive bandwidth method to fit the joint PDF P(error, dist), then
numerically integrate and divide to get the marginal, conditonal, and
finally the conditional CDF.

This all works (seemingly) well. Next comes the plotting.